+++ pkg/parser/parser.go
@@ -157,12 +157,12 @@ func (e *expr) Metrics() []MetricRequest {
 			var r2 []MetricRequest
 			for _, v := range r {
 				for i := int64(start); i < int64(end); i++ {
-					fromNew := int64(v.From) + i*int64(offs)
-					untilNew := int64(v.Until) + i*int64(offs)
+					fromNew := v.From + i*int64(offs)
+					untilNew := v.Until + i*int64(offs)
 					r2 = append(r2, MetricRequest{
 						Metric: v.Metric,
-						From:   int64(fromNew),
-						Until:  int64(untilNew),
+						From:   fromNew,
+						Until:  untilNew,
 					})
 				}
 			}
@@ -180,8 +180,8 @@ func (e *expr) Metrics() []MetricRequest {
 					return nil
 				}
 				for i := range r {
-					fromNew := int64(r[i].From) - int64(offs)
-					r[i].From = int64(fromNew)
+					fromNew := r[i].From - int64(offs)
+					r[i].From = fromNew
 				}
 			}
 		}
